String resources are updated. TR Language translations are updated. CapacitySelectionFragment is merged with CapacityPriceSelectionFragment.
